tools: move memory classes to tools.memories
