+++ test/test_tools.py
@@ -118,54 +118,3 @@ def test_chain_loaders(configfactory):
     results = loader(settings)
 
     assert results == [re_numeric, re_text]
-
-
-def test_sopel_identifier_memory_str():
-    user = tools.Identifier('Exirel')
-    memory = tools.SopelIdentifierMemory()
-    test_value = 'king'
-
-    memory['Exirel'] = test_value
-    assert user in memory
-    assert 'Exirel' in memory
-    assert 'exirel' in memory
-    assert 'exi' not in memory
-    assert '#channel' not in memory
-
-    assert memory[user] == test_value
-    assert memory['Exirel'] == test_value
-    assert memory['exirel'] == test_value
-
-
-def test_sopel_identifier_memory_id():
-    user = tools.Identifier('Exirel')
-    memory = tools.SopelIdentifierMemory()
-    test_value = 'king'
-
-    memory[user] = test_value
-    assert user in memory
-    assert 'Exirel' in memory
-    assert 'exirel' in memory
-    assert 'exi' not in memory
-    assert '#channel' not in memory
-
-    assert memory[user] == test_value
-    assert memory['Exirel'] == test_value
-    assert memory['exirel'] == test_value
-
-
-def test_sopel_identifier_memory_channel_str():
-    channel = tools.Identifier('#adminchannel')
-    memory = tools.SopelIdentifierMemory()
-    test_value = 'perfect'
-
-    memory['#adminchannel'] = test_value
-    assert channel in memory
-    assert '#adminchannel' in memory
-    assert '#AdminChannel' in memory
-    assert 'adminchannel' not in memory
-    assert 'Exirel' not in memory
-
-    assert memory[channel] == test_value
-    assert memory['#adminchannel'] == test_value
-    assert memory['#AdminChannel'] == test_value
